organized structure of files
