+++ scripts/neuron_prinz.jl
@@ -7,17 +7,17 @@ using NeuronBuilder, ModelingToolkit, OrdinaryDiffEq, Plots
 const Area = 0.0628 # Prinz/Liu 0.0628 mm2
 const Cm = 10.0 # specific capacitance cₘ is a biological constant (around) 10 nF/mm^2
 
-Prinz2_conv = Cm / Area
+Prinz_conv = Cm / Area
 
-channels = [Prinz2.Na(100 * Prinz2_conv), Prinz2.CaS(6 * Prinz2_conv), Prinz2.CaT(2.5 * Prinz2_conv), Prinz2.H(0.01 * Prinz2_conv),
-    Prinz2.Ka(50 * Prinz2_conv), Prinz2.KCa(5 * Prinz2_conv), Prinz2.Kdr(100 * Prinz2_conv), Prinz2.leak(0.0)]
+channels = [Prinz.Na(100 * Prinz_conv), Prinz.CaS(6 * Prinz_conv), Prinz.CaT(2.5 * Prinz_conv), Prinz.H(0.01 * Prinz_conv),
+    Prinz.Ka(50 * Prinz_conv), Prinz.KCa(5 * Prinz_conv), Prinz.Kdr(100 * Prinz_conv), Prinz.leak(0.0)]
 
 τCa = 200.0
 Ca∞ = 0.05
 
 defaults = Dict(Voltage => BasicVoltageDynamics(),
-    Calcium => PrinzCalciumDynamics(τCa, Ca∞, 14.96 * 0.0628),
-    Reversal{Calcium} => PrinzCaReversalDynamics())
+    Calcium => Prinz.CalciumDynamics(τCa, Ca∞, 14.96 * 0.0628),
+    Reversal{Calcium} => Prinz.CaReversalDynamics())
 
 somatic_parameters = Dict(
     Reversal{Sodium} => 50.0,
