+++ src/assets/Prinz/Prinz.jl
@@ -3,11 +3,12 @@ module Prinz
 using ModelingToolkit
 using ..NeuronBuilder
 
-import ..channel_dynamics, ..ionic_current, ..calcium_current, ..external_params
+import ..get_parameters, ..get_states, ..default_params, ..default_states
 
 include("channels.jl")
+include("calc_dynamics.jl")
 
-export channel_dynamics, ionic_current, calcium_current, external_params
+export get_parameters, get_states, default_params, default_states
 
 end
