--- src/assets/Prinz2/Prinz2.jl
@@ -1,13 +1,0 @@
-module Prinz2
-
-using ModelingToolkit
-using ..NeuronBuilder
-
-import ..channel_dynamics, ..ionic_current, ..calcium_current, ..external_params
-
-include("channels.jl")
-
-export channel_dynamics, ionic_current, calcium_current, external_params
-
-end
-
