+++ src/build_network.jl
@@ -1,9 +1,33 @@
-"""
-1. number of hooks for each neuron. This is in-degree of neuron
-"""
+function add_connection(group, pre, post, syn::EmptyConnection; kwargs...)
+    return group
+end
 
+function add_connection(group, pre_n, post_n, syn::Synapse; name=ModelingToolkit.getname(group), i=1)
+    pre = pre_n.sys
+    post = post_n.sys
+    prename, postname = ModelingToolkit.getname.([pre, post])
+    synapse_sys = syn(; name=Symbol(prename, :to, postname, get_name(syn)))
 
+    oldeqs = ModelingToolkit.get_eqs(group)
+    neweqs = [
+        get_pre(syn, pre) ~ my_pre(synapse_sys),
+        get_post(syn, post) ~ my_post(synapse_sys),
+        syn_current(syn, synapse_sys.sys) ~ getproperty(post, Symbol(post_connector(syn), i))
+    ]
 
+    eqs = cat(oldeqs, neweqs, dims=1)
+    systems = cat(ModelingToolkit.get_systems(group), synapse_sys.sys, dims=1)
+
+    return connected = ODESystem(eqs, t, [], [];
+        name=name,
+        systems=systems)
+end
+
+function in_degree(edge_fn, num_nodes, j)
+    sum(1:num_nodes) do i
+        length(edge_fn(i, j)) - (typeof(edge_fn(i, j)) == Tuple{EmptyConnection})
+    end
+end
 
 function build_network(nodefn, edgefn, num_nodes; name=:network)
     eqs = Array{Equation,1}(undef, 0)
@@ -20,14 +44,3 @@ function build_network(nodefn, edgefn, num_nodes; name=:network)
     end
     return network |> structural_simplify
 end
-
-
-function in_degree(edge_fn, num_nodes, j)
-    sum(1:num_nodes) do i
-        length(edge_fn(i, j)) - (typeof(edge_fn(i, j)) == Tuple{EmptyConnection})
-    end
-end
-# if typeof(syn) !== EmptyConnection
-#     println(typeof(syn), i, j)
-#     println(ModelingToolkit.getname(neurons[i].ODESystem), " is pre and ", ModelingToolkit.getname(neurons[j].ODESystem), " is post \n")
-# end
